backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.4
defusedxml==0.8.0rc2
Django==4.2.3
django-celery-results==2.5.1
django-cors-headers==4.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-writable-nested==0.7.0
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
kombu==5.3.2
oauthlib==3.2.2
outcome==1.3.0
packaging==23.2
Pillow==10.0.0
prompt-toolkit==3.0.39
pycparser==2.21
PyJWT==2.7.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8
webdriver-manager==4.0.1
wsproto==1.2.0
